final v7

+++ Servlet&JSP/Demo1/Bus/edit_bus2.jsp
@@ -7,12 +7,27 @@
 <title>view_bus_jsp</title>
 </head>
 <body>
-  
 <%@ page import = "java.sql.*" %>
 <%@ page import = "javax.sql.*" %>
 <%@ page import = "java.util.*" %>
 <%@ page import = "javax.servlet.*"%>
 <%@ page import="java.io.*" %>
+
+
+<%
+  //response.setHeader("Cache-Control", "no-cache, no-store, must-revalidate"); 
+
+  out.println("<div style='position: absolute; top: 0; right: 1; width: 100px; text-align:right;'><br><br><br><br><br><br><br>");
+  out.println("<a href='../logout_admin.jsp' style='font-size:20px;'>Logout</a><br>");
+  out.println("</div>");
+// response.setHeader("Pragma", "no-cache"); 
+// response.setDateHeader("Expires", 0);
+
+if(session.getAttribute("userid")==null)
+{
+    response.sendRedirect("../admin.jsp");
+}
+%>
 
 <%
 try
@@ -31,7 +46,6 @@ try
     request.getRequestDispatcher("edit_bus1.jsp").forward(request, response);
   }
 
-    
  Class.forName("org.postgresql.Driver").newInstance();
  Connection c = DriverManager.getConnection("jdbc:postgresql://localhost:5432/postgres","postgres", "1234");
  c.setAutoCommit(false);
@@ -39,26 +53,37 @@ try
 
 Statement stmt = c.createStatement();
 
-String sql="delete from bus_details where bus_id = " + bus_no;  
-stmt.executeUpdate(sql);
+ ResultSet rs = stmt.executeQuery( "SELECT * FROM bus_details where bus_id='"+bus_no+"'");
+if(rs.next())
+{
 
-out.println("Bus deleted successfully");
+  String sql="delete from bus_details where bus_id = " + bus_no;  
+  stmt.executeUpdate(sql);
 
+  //out.println("Bus deleted successfully");
 
-String sql1 = "INSERT INTO Bus_Details (BUS_ID,BUS_NAME,TOTAL_SEATS,AC_or_NON_AC,DEPARTURE,DESTINATION)"
-  + "VALUES ('"+bus_no+"','"+bus_name+"','"+seats+"','"+sleeper+"','"+dep+"','"+des+"');";
+  String sql1 = "INSERT INTO Bus_Details (BUS_ID,BUS_NAME,TOTAL_SEATS,AC_or_NON_AC,DEPARTURE,DESTINATION)"
+    + "VALUES ('"+bus_no+"','"+bus_name+"','"+seats+"','"+sleeper+"','"+dep+"','"+des+"');";
 
-stmt.executeUpdate(sql1);
+  stmt.executeUpdate(sql1);
 
-out.println("<br>Bus Added successfully<br>");
-stmt.close();
-c.commit();
-c.close();
+  //out.println("<br>Bus Added successfully<br>");
+  stmt.close();
+  c.commit();
+  c.close();
 
-  String message = "Bus added Successfully!";
-  request.setAttribute("message", message);
-  request.getRequestDispatcher("edit_bus1.jsp").forward(request, response);
+    String message = "Bus Edited Successfully!";
+    request.setAttribute("message", message);
+    request.getRequestDispatcher("edit_bus1.jsp").forward(request, response);
 
+  }
+else
+{
+    String message = "No bus with this number exists!";
+    request.setAttribute("message", message);
+    request.getRequestDispatcher("edit_bus1.jsp").forward(request, response);
+
+}
 }
 
 catch(Exception e){
